make --watch work with vim (#24)
Vim appears to move/rename the original file on save, resulting
in a  RENAME event from fsnotify. Reacting to this by re-adding
the original file name also catches two CHMOD (but no WRITE).

Not sure what's going on, but re-adding on RENAME and reloading
on CHMOD seems to do the trick.