Add New Font (#72)
* Use upstream go-bdf library instead of our fork.

The zachomedia/go-bdf has our changes merged and it has fixes to the
bounds of the glyphs, which renders fonts much more nicely using pixlet.
For example, a colon now has the appropriate spacing.

* Added font embedding to the Makefile.

This commit adds the font embedding utility to the make file so that we
can embed new fonts on build.

* Added a font preview example.

This commit adds a preview to examples to preview new fonts in the repo.

* Added tom-thumb font.

This commit adds the Tom Thumb font found here: https://robey.lag.net/2010/01/23/tiny-monospace-font.html

This font is supy tiny and allows for packing some serious information
onto a display.

* Fixed false comment.

The all uppercase comment can be true, but we'd need to modify this
font. I honestly think we should add a tom-thumb-upper that has all
lowercase characters removed and the height set to 5.