Updated our fork of starlib. (#75)
This is getting pretty unwieldy, but we have a fix. This change gets the
fix we need right now, and Rohan is cutting a PR with his changes that
have not yet been upstreamed. Once that's merged, we can use upstream,
but it will require time and math module changes.