Upgrade to mainline starlib. (#76)
This commit removes the override to use the Tidbyt fork of starlib.
Addtionally, it upgrades past v0.5.0 which includes breaking changes to
time, math, and json. I am using the latest commit off of the main
branch because we need the most recent change to get rid of our fork. We
can move to a release once one is cut for starlib with the change we
need.