feat(arm64-support): add arm64 to goreleaser config
this adds arm64 to the list of supported go architectures in the
goreleaser configuration.

it also updates the fetch-artifacts.sh script to copy the linux and
darwin arm64 artifacts. i'm not entirely sure how this fits into the
buildkite CI process, but it seems like it might need to be updated.

i'm able to run the goreleaser command locally and see it generate all
the proper files.

one interesting and important note about building the render module
while running on macos, i did need to set up the CPATH and LIBRARY_PATH
as described here https://github.com/tidbyt/pixlet/issues/39#issuecomment-900727783