feat: Overhaul Hot Reloads (#86)
* Added new server packages.

This commit open sources a few modules that were internal to Tidbyt and
refactors existing server code into a loader and watcher module. Some
big changes in the refactor include directory watching inside of watcher
to ensure we get every update and removing sync.Mutex in favor of
channels.

* Refactored server to use new packages.

This commit refactors the server code to use our new packages.

* Update build to go1.16 to match go.mod.

This commit updates the test actions to use go 1.16.12 to match the
go.mod and to be able to support go embpedding.