fix: Standardizes the way we build (#87)
* Remove font embedding from build.

I added this when I was doing a lot of font work, but now that I'm not
it really is annoying to generate fonts each time. This commit makes it
so the only target for build is a clean.

* Remove build directory in make.

This is unncessary and I feel bad for adding it. I've removed it so that
a build places the binary in the root.

* Updated build to use make.

This commit updates the build to use make so there is one way to build
things.

* Updated build docs.

This commit updates the build docs to use the one true way to build this
repository.