Added release scripts.
This commit adds release scripts that can cross compile with CGO support
for MacOS and Linux for both x86_64 and arm64 based CPU architectures.
It assumes that each build is being build on the target platform though.
Cross compiling the world on Linux would be ideal, but it would require
all of those toolchains to be setup on our build hosts. Doable, but I'll
leave it for a future excercise.