feat(release): Move release to GitHub Actions.
This commit moves our release process to GitHub actions. The way this
used to work was through our internal build system which makes it really
hard to give any visibility to anyone else. In addition, we were hitting
the limits of our local macOS build system and needed an upgraded agent.