feat: Open Source Schema (#117)
* Open source starlarkutils.

This commit open sources starlarkutils that attaches a context to a
running starlark thread. We need it to call schema handlers with a
context.

* Open source schema package.

This commit adds the schema package to the pixlet repo. Schema has an
interface similar to render.

* Added schema methods to applet.

This commit adds the necessary methods to runtime.Applet to be able to
manage schema.

* Added schema documentation.

This commit adds schema documentation.