schema: Strongly type defaults and config values
Support config values of type other than string. Previously we just
assumed all values were strings, which is kind of crappy for things like
toggles where we were checking stuff like `config.get("X") == "true"`.

Instead, we now check against the schema and provide a value of the
correct type to Starlark. We still serialize everything as a string for
storage, but we deserialize it to the appropriate type.