sunrise: Add sunrise module to pixlet. (#124)
* sunrise: Add sunrise module to pixlet.

This commit adds the go-sunrise module to pixlet so that apps can make
use of sunrise/sunset times. This has been a popular use case and
hopefully this module will make it easier for more folks to take
advantage of this information.

* Fix accidental change

* sunrise: Add better support for no sunrise/sunset.

This commit adds the None return type for an empty time zone and added
it to the example.