Revert strongly-typed values for `config.get`
Starting with #120, `config.get` would return values of different types
based on the type of the underlying schema field. For example, for
toggle fields, the returned value would be a boolean instead of a string.

Unfortunatley, there are multiple edge cases where this doesn't work.
For example, if a schema field gets deleted but an installation still
has a value for it, then we no longer know what type to return.

On top of that, for generated fields, we don't have an easy way to tell
what type they are. Generated fields aren't currently implemented in
Pixlet, but they are available to Tidbyt-internal apps and will be in
Pixlet soon.