Allow apps to declare secret values (#129)
* Allow apps to declare secret values

Add support for a `secrets` dictionary in `Schema`. Apps can declare
secrets by providing an encrypted value that can only be decrypted by a
private key.

Each `runtime.Applet` now accepts a `SecretDecryptionKey`. If a key is
provided, the runtime will use it to decrypt all of an app's secrets
when loading the Starlark script.

The decrypted values are provided to Starlark via the `config` object.
During development, a value can be provided via `pixlet render`, or a
querystring when using `pixlet serve`.

* Add `pixlet encrypt` command

The encrypt command encrypts secret values so that they can only be
decrypted by a specific app running on the Tidbyt community servers.