add max_age to Root widget (#143)
When a Tidbyt device fails to receive new screen data for an
installation - e.g. due to a connectivity issue, or a bug in an
applet - it will keep displaying the most recently received graphic.
That's not a desirable behavior for something like a clock, where
stale data could cause someone to be late for a meeting, or get
out of bed earlier than needed.

The max_age attribute will be passed on to the Tidbyt device,
allowing it to expire stale screen data.