schema: Fix schema options. (#145)
This commit fixes a few things with schema options that were wrong. I
renamed "text" to "display" since I found it more descriptive. The issue
is the object wasn't renamed when it was injected into config. In
addition, I didn't consider that we could get an "option" back from
schema handlers when I moved to the syntactic sugar version of schema.