feat: Devtools (#155)
* devtools: Add React based devtools.

This commit adds a React project containing a brand new browser
interface for pixlet.

* devtools: Update browser to support React devtools.

This commit updates the browser module to be able to support the newly
added React frontend.

* devtools: Updated GitHub workflows.

This commit updates the GitHub workflows to be able to build the React
frontend.

* devtools: Move goreleaser to use ./out.

This commit updates goreleaser to not clash with the dist directory that
is not being used with the React frontend.

* devtools: Updated the build instructions.

This commit updates the build readme so that folks know how to build the
new frontend.

* devtools: Fix build.

The frontend build was broken due to a name collision on theme. It works
on case sensitive filesystems but not on case insensitive filesystems.

* devtools: Handle no default when optional.

This commit fixes a load issue when defaults are optional, which is only
true for a few components.

* devtools: Add scopes for OAuth2.

This commit provides scopes for OAuth2.

* devtools: Swap token for code.

This commit updates the OAuth2 handler to get a code instead of a token
to be consistent with our mobile app.

* devtools: Fix toggle.

This commit fixes the default for toggle.

* devtools: Fix go tests.

This commit fixes go tests by renaming the go template file.

* devtools: Fix issue with empty screens.

This commit fixes an issue where empty screens were not being displayed
properly.