Fix hash calculation for widegts that have NaN values
The `Plot` widget now supports `NaN` values for some attributes. This
caused a problem when calculating hash, since we were getting the hash
by marshaling the render tree to JSON. Officially JSON doesn't support
`NaN` or inifinities.

Instead switch to marshaling to MessagePack. This is faster and supports
a wider range of data types.