Marquee: Add fix for alignment of non-scrolling child (#298)
This allows for alignment of the child when it fits fully, this is especially useful for when text changes and may or may not fit.

It should work with other child objects, but unit tests should be used to confirm the intended behavior.

Added a unit test for "center" and "end" alignment. Basically just copied the TestMarqueeNoScrollHorizontal and lengthened width and height of the Marquee and added alignment.