Print profile directly
Hooray, it is possible!

Rather than writing to a gzipped file and then having to install and run a separate tool, we can do it all in one.

This is all using the public API from the pprof github repo.

The pprof tool wants to run in REPL mode. So, on the first loop we give it the command we want and it prints the output. Then on the second loop, we exit successfully.