Limit number of parallel frames being painted
In #67, we started painting frames in parallel instead of serially, in
order to fully utilize available CPU.

But we did this without any limit. So if a `Root` had hundreds of frames,
we would attempt to paint all of them in parallel. This would consume
huge amounts of memory by allocating hundreds or thousands of image
buffers.

Instead, we now limit parallelism to the number of CPU's.