Limit the maximum number of frames for an app
Apps can return an arbitrarily huge number of frames in their render
tree. When painting, we keep all frames in memory to assemble an
animated WebP or GIF, so this can get out of control.

Instead, set a default maximum of 2000 frames per app. Any frames past
the 2000th frame will be discarded.