Add day_of_week to humanize module (#401)
I've had a couple of apps where I've needed to know the day of the week and ended up implementing Zeller's congruence
https://en.wikipedia.org/wiki/Zeller%27s_congruence

Seems like something that should be easy to get from a library. Someone's also asked for it on the forum
https://discuss.tidbyt.com/t/how-to-derive-day-of-week/2555/2.

Only after doing this have I just realised that it's basically possible using time.format("Mon") or time.format("Monday"). But a) if it wasn't obvious to me then it's probably not obvious to someone else, and b) I don't think you should have to mess about with strings to do this kind of maths.