Add `pixlet login` for logging in via UI
Add `pixlet login`, which launches login.tidbyt.com in a web browser
for the user to login. It also starts an HTTP server in the background
to accept the OAuth 2.0 callback from login.tidbyt.com, and stores the
API token to a config file.

We then use the token in `pixlet push` so that the user doesn't have to
provide an API token explicitly, just the device ID.