build: Fix gitignore. (#465)
This commit should actually resolve the dirty state issue with
goreleaser. It wasn't that the directory was left behind, it was that
the directory existed when goreleaser started. The reason the others
were ok was because these directories only have one file. However, we
never ignored the `.exe` version of pixlet.