build: Fix macOS intel based release.
This commit fixes the macOS intel based home brew release. The overview
is that GoReleaser sets these variables and expects them all to exist.
However, when we switch to the prebuilt mode, most assumptions
GoReleaser makes goes out the window. The docs do say that these
variables will not be set and we have to define them ourselves, but the
list of targets felt just fine at the time. The issue is, the hombrew
logic expects these variables to exist and not just a list of targets. I
added back the old format and added in the defaults of the other
variables that are no longer set.