Return nodes from XPath queries (#424)
* Refactor XPath into module

It seems much more like the other modules like random than a core part of running an applet.

* Move test for XPath module

Again, this is how other modules do it. This doesn't seem to be a core part of an applet, and moving it makes it easier to extend.

* Return nodes from XPath queries

This would have helped with the national rail app.

I was working with an API that returns XML. The response contained a list of items and I wanted to extract a few values from each item.

Say I am dealing with item `/foo/bar/baz[0]` and I want fields `qux` and `xyz` with in it.

Currently, I have to repeat the absolute path in every query -- `/foo/bar/baz[0]/qux` and `/foo/bar/baz[0]/xyz`. This also presumably means we have to do more work for each query because we walk the full length from the root to the leaf each time.

With this PR, it's possible to instead do one query first to get the node for the item -- `/foo/bar/baz[0]/` -- and then do smaller queries on that node -- simply `/qux` and `/xyz`.

* Add documentation