bugfix: Fix widget codegen (#539)
* codegen: Update codgen with a Make command.

This commit updates the code generation we use for generating docs and
widgets to be able to be run from the root directory. In addition, it
adds a Make command so it's more obvious how to run it.

* codgen: Fix stale/out of sync docs.

This commit fixes the stale and out of sync documentation. There was
some amount of drift between the code generation and what existed in our
documentation. This commit fixes the source so that `make widgets` can
be run safely.

* codegen: Remove example files.

We don't use the example files anywhere, and it's causing issues with
buildifier.