codegen: Run gofmt as part of code generation (#540)
* codegen: Add formatting to gen commands.

This commit adds gofmt to run as part of the code generation. A cleaner
fix would be to fix all of the templates in code generation, but it
would only solve the problem if the templates never changed.

* codgen: Run gofmt.

This commit runs `gofmt -s -w ./` to format and simplify go code in this
project.