feat: Add native linter and formatter (#522)
* feat: Add lint command to pixlet.

This commit adds a lint command to pixlet that wraps buildifier. Right
now, it works as a drop in replacement for buildifier, but it's not very
user friendly from a pixlet perspective. I'll follow up with another
commit to make this easier to use from a pixlet lens.

* feat: Break apart lint and format.

This commit splits linting and formatting into two commands and cleans
up all of the buildifier flags so that it's easier to use in pixlet.

* Clean up lint command and interface.

This commit cleans up the lint command, help, and examples to behave as
discussed in Discord.

* Clean up format command and interface.

This commit cleans up the format command and interface to behave has we
disscussed in Discord.

* Update format to output.

When I was typing these commands, I felt the urge to add `-o json`
instead of --format. Given it felt more natural to me, I updated the
command arg.

* Fix warnings list to include all warnings.

This commit updates the linter to include all warnings. I originally
left this out by mistake, so this is a bug fix. We may consider a
different list here in the future but it will require more testing.

* Drop output for format.

The output format doesn't really make sense for formatting. As a pixlet
user, I only care about formatting or a diff. The json format only
spits out the files processed, which I don't really care about at this
time.

* Fix formatting when fix flag is set.

When linting, we need to fix the formatting in order to also fix the
resolvable lint issues. This commit ensures we do both if the --fix flag
is set.

* Set default warning list.

This commit sets up a default warning list with a disabledWarnings map
to be able to control which warnings we enable.