Disable some lint warnings. (#560)
This commit disables lint warnings for unsorted dictionaries and naming
conventions. For dicts, the issue is there are a lot of dictionaries
that are logcially sorted in code and this would rewrite all of them.
For the naming, we have 2.8K naming convention violations in the
community repo so it would be really hard to enforce the other linters
if we didn't fix those first.