tools: Update create command to support app types.
This commit updates the code generation for new apps to be aware of the
repo a user is in and be able to make decisions accordingly. If the user
is in the community repo, it will create the app in the appropriate
location. Otherwise, it will create the app in the local directory. I
also added support for our internal repo so that I can reuse this
internally :D.