tools: Update the default app.
This commit updates the default app to include a schema field. The
default app should be formatted, lint, and be ready to publish. At the
same time, I want the default app to showcase the things you would need
to produce an app. So something like `config` gets weird if it's not
used because the linter will complain. But it would be weird to not have
it because a user might get confused.