feat: Add create command (#571)
* tools: Create tools package.

This commit creates the tools package with the addition of git based
functions for determining if you're in a given repo or not.

* tools: Add in community repo tools.

This commit migrates the community repo tools package to exist in the
pixlet repo instead. This is a stepping stone to enable a `pixlet
create` command to create apps using pixlet.

* tools: Add create command from community repo.

This commit adds the create command from the community repo. As of right
now, it does a quick check to determine if the user is running from the
root of the community repo. More work will be done to be able to run
create app from anywhere.

* tools: Update create command to support app types.

This commit updates the code generation for new apps to be aware of the
repo a user is in and be able to make decisions accordingly. If the user
is in the community repo, it will create the app in the appropriate
location. Otherwise, it will create the app in the local directory. I
also added support for our internal repo so that I can reuse this
internally :D.

* tools: Add more creation information.

This commit adds the absolute path to where the app was created as well
as a command using the relative path to start the app.

* tools: Update the default app.

This commit updates the default app to include a schema field. The
default app should be formatted, lint, and be ready to publish. At the
same time, I want the default app to showcase the things you would need
to produce an app. So something like `config` gets weird if it's not
used because the linter will complain. But it would be weird to not have
it because a user might get confused.

* tools: Update help text.

This commit updates the help text to make it more useful as a general
command.