feat: Standardize error handling.
This commit standardizes error handling to return an error from commands
using RunE. In addition, it silences the usage on error so it's only our
error message printed. This has a few benefits, but the one I'm after is
being able to call commands from another command. For example, a `pixlet
check` command that can call other commands and handle their output.