feat: Add app publish check command (#574)
* render: Add the ability to silence print statements.

This commit adds a `--silent` flag to silence print statements in Tidbyt
apps.

* lint: Add an option to turn output off.

This commit adds an option to turn off the output. This is useful if we
only care about if an app is linted or not, but don't care about the
output.

* feat: Add check command.

This commit adds a check command to be able to ensure an app is ready
for publishing.