lint: Run the linter twice with --fix.
This commit resolves an issue where the --fix flag would cause no output
and a zero exit code if there were issues resolved, even if it could not
resolve all issues. To get around that, this change runs the linter a
second time without fix to ensure the exit is still 0 or print the
unresolvable errors.