feat: Community Target Determination (#576)
* community: Create target determinator command.

This commit adds a base target determinator command that is cabable of
providing the changed files between two commits.

* community: Update target determinator to use tree diff.

This commit updates the method for finding files to use built-ins for
diffing the tree. My original method of going through commits linearly
wouldn't work in all cases and would be error prone.