fix: Clean up from manual testing (#577)
* check: Fix issue with directory and no rflag.

This commit resolves an issues where a directory without a recursive
flag would cause non standard errors by adding validation.

* lint: Run the linter twice with --fix.

This commit resolves an issue where the --fix flag would cause no output
and a zero exit code if there were issues resolved, even if it could not
resolve all issues. To get around that, this change runs the linter a
second time without fix to ensure the exit is still 0 or print the
unresolvable errors.

* lint: Ignore docstring warning.

I wasn't sure if we were going to use the docstrings in the community
repo. I've ultimately decided against it and am disabling the linter
since it's not necessary.