feat: Add YAML support for app manifests (#580)
* tools: Add YAML support to manifests.

This commit adds struct tags for yaml and provides the ability to load
and write a file using the reader/writer interface.

* tools: Migrate manifest to a top level pkg.

This comiit migrates the manifest to a top level package. It's going to
be used in our internal repo, which makes tools a weird spot for it.

* tools: Add manifest generation.

This commit adds in the YAML based manifest generation.

* tools: Remove apps.go and manifest.go generation.

This commit removes the go code generation that we will no longer need.

* community: Add validate-manifest command.

This commit adds a command to the community sub-command that will take a
manifest file as an argument and validate that it's correct.

* community: Add create-manifest command.

This commit adds a create manifest command that can create a manifest if
you already have an app. The benefit is for folks who might be looking
to publish their app but don't need to create the starlark over again.

* check: Add manifest validation.

This commit adds manifest validation and existance checks to the pixlet
check command to ensure an app both has a manifest and it's valid.