Create InputSlider widget
Latitude and longitude are numbers with a bounded range, so a slider makes sense. But I would often want to test a specific location so I also want to be able to type in the number.

This creates a linked slider and numerical input. Changing one changes the other, and both have the same limits.

Largely based on an example on the MUI site.
https://mui.com/material-ui/react-slider/#InputSlider.js