Create skeleton LocationForm control
Nothing interactive yet, but this creates a form that lets the user manually input a specific location.

These are the four fields in the location object used by the config. Latitude and longitude are bounded by nature. Location name is an arbitrary string.

Timezones are populated based on the user's system environment. I assume that makes sense, but I may be wrong given discord chatter about Windows users needing the right library for things to work.

Can't call it Location because that's a reserved keyword and it wouldn't compile if called that.