feat: Icon Validation (#590)
* community: Add icon validation.

This commit adds the icon validation from the community repo to ensure
the icons selected are available in our backend.

* community: Add command to list icons.

This commit adds a command to list valid icons. This will be useful for
the check command so users can find valid icons.

* check: Add check for icon validation.

This commit adds the icon validation into the pixlet check command.