fix: Community Integration Bugs (#595)
* fix: Resolve an issue with manifest file names.

This commit resolves an issue where the manifest file was not being
written to the correct location and was overwriting the starlark source,
but only in the community repo.

* fix: Make the manifest filename a constant.

This commit removes the ability to have a .yml ending switches all uses
to be a constant. This will significantly reduce headaches down the
line given there is only one file name.

* community: Add load app command.

This commit adds a load app command so that we can ensure that an app
can load, even if it cannot render.

* fix: Update checks to match community.

This commit makes some of the checks a bit less aggressive with a TODO
on how to make it more agressive in the future. The reason being, I'd
like to get this enabled today as a drop in replacement rather then be
more strict.