Embed timezone database in pixlet
I think this might fix https://github.com/tidbyt/pixlet/issues/551

Importing this into the binary's main embeds a copy of the timezone database. That should put an end to Windows users having to do extra steps to work with most of the community apps. Cost is a slightly larger binary.

https://pkg.go.dev/time/tzdata

I don't have a Windows machine to try it out, but the test will fail if America/Los_Angeles is an unknown timezone.