Embed timezone database in Windows build (#616)
I think this might fix #551

Alternative to https://github.com/tidbyt/pixlet/pull/615

Windows users keep running into problems because their system doesn't have the timezone database installed. Other platforms have it.

Per https://pkg.go.dev/time/tzdata, building with this flag will embed a copy of the database into the binary, so that the lookup just works out of the box.

I don't have a windows machine to try it out, but I added a test that should fail on Windows if the database is not present.