Support schema.LocationBased in pixlet serve (#586)
Similar to https://github.com/tidbyt/pixlet/pull/585 but more important.

This adds a LocationBased widget that is very similar to LocationForm. However, when the location is set, it calls the relevant handler to get a list of options, and then displays them in a dropdown menu.

The chosen option is added to the URL and persisted on page reload. However, the location isn't. It could be done by adding the location to the widget's value object, but I don't want to introduce something that isn't supported in production. As far as I know, only the selected option value should be available to the starlark script.

I wonder if it's possible to add a button to dismiss snackbars. I often got a permanent red box because the handler didn't like receiving a '55.' because it doesn't parse as an int, but I was mid-typing!