feat: App bundles (#665)
* feat: Add app bundles module.

This commit adds a module that can create and extract bundles of apps.
This will be useful for being able to create portable apps in a
standardized format.

* feat: Add bundle CLI command.

This commit adds a CLI command to produce a bundle.

* feat: Add bundle upload and deploy commands.

This commit adds bundle upload and deploy commands. They are seperate so
that a rollback can occur without needing to have the bundle on hand.

* Remove todos

* Fix bugs

* fix: Update auth for app bundles.

This commit udpates the auth story for app bundles. It adds a new
set-auth command that is capable of setting the configration file with a
JSON encoded token.

* Fix windows test

* Fix windows test again