Pixlet Create: Add commands to specify path folder for new app (locally)
The behavior for if in tidbyt or community repo should nto be changed. Had to add a few if/else statements to generator.go since appsDir is set to be a constant and was not sure if that would break things internally for tidbyt, so made the changes to fix locally. Also did not want to add more info to the generator structure or app structure