feat: HTTP Caching Client (#766)
* http: Vendor starlib http client.

This commit vendors in the starlib http client and adds in Tidbyt
specific headers. These are used by our caching client to make decisions
about how to cache HTTP requests.

* http: Add caching client that mimics internal proxy.

This commit adds a caching HTTP client that mimics our internal proxy.

* http: Wire in caching http client.

This commit wires in the http client into all of the places that need a
new app runtime.