Expose widget frame count
Saw this issue asking for a way to get the frame count for an Animation or Marquee widget https://github.com/tidbyt/pixlet/issues/823

Looks like that's defined for all widgets (and is 1 in case it's not animated). This just updates the template for the render module to expose it as a readonly method, following the way size() is available for Text widgets.

Added test cases. Also checked that it gives a sensible answer for marquees in particular, which gets longer as the text gets longer.