bundle: Add WithoutRuntime write option (#1051)
WithoutRuntime is a WriteOption that can be used to write the bundle without
using the runtime to determine the files to include in the bundle. Instead,
all files in the source FS will be included in the bundle.

This is useful when writing a bundle that is known not to contain any
unnecessary files, when loading and rewriting a bundle that was already
tree-shaken, or when loading the entire runtime is not possible for
performance or security reasons.