Serve GIFs as well as WebP
Saw https://github.com/tidbyt/pixlet/issues/1005 and thought it was a fun little challenge.

Tested that this worked locally for:
 - Previewing with WebP (default)
 - Previewing with GIF (passing --gif flag)
 - Updating config using either format
 - Exporting image using either format

Inspected the images shown and confirmed that it does generate a WebP or GIF file as appropriate, and that the src element is set to the appropriate image type.