Add ID field for sounds
Both the path and the contents of a notification sound could conceivably
change without the sound being semantically different.

For example, `ding.mp3` could get renamed to `ding1.mp3`. Or a developer
may re-encode `ding.mp3` to make it sound better.

In both cases, we need some sort of ID to keep a handle on the sound and
know what to play.